+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - JavaScript unit tests using Jest
 - ESLint with Prettier
 - Templated `robots.txt` that can be configured per deployment environment
+- Image optimisation task
 
 ### Changed
 - Site build now uses Metalsmith